Complete

+++ src/main.h
@@ -18,7 +18,7 @@ typedef struct
     uint16_t HighVoltage;
     uint16_t LowVoltage;
     uint16_t HighTemp;
-    uint16_t HighImp;
+    uint32_t HighImp;
     uint16_t alarmSetStatus;
     char deviceName[20];
 } nvsSystemSet;
